chore: 파일구조 분리
