+++ src/Class/WinningNumbers.js
@@ -1,0 +1,11 @@
+class WinningNumbers {
+  #winningNumber;
+  #bonusWinningNumber;
+
+  construnctor() {
+    this.#winningNumber = "";
+    this.#bonusWinningNumber = "";
+  }
+}
+
+export default WinningNumbers;
