+++ src/constants/lotto.js
@@ -1,0 +1,3 @@
+export const LOTTO_PRICE = 1000;
+export const LOTTO_COUNT = 6;
+export const LOTTO_NUMBERS = Array.from({ length: 45 }, (_, i) => i + 1);
